Fixing regression introduced with JSON output.
If the client did not send a Content-Type header with their request the code pulling it from the dict will fail with a KeyError.
To remedy this the code was changed to use request.headers.get() to defaul to None when the header is not present rather than failing
with a KeyError.